backend_django/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
Pillow==10.4.0
gunicorn==23.0.0
whitenoise==6.7.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
